tests/wpt/web-platform-tests/html/interaction/focus/focusgroup/tentative/forward-navigation/does-not-move-when-outside-focusgroup.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML Test: focusgroup - Focus does not move when initially set on an element not included in the focusgroup.</title>
<link rel="author" title="Microsoft" href="http://www.microsoft.com/">
<link rel="help" href="https://github.com/MicrosoftEdge/MSEdgeExplainers/blob/main/Focusgroup/explainer.md">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="../resources/focusgroup-utils.js"></script>

<span id=out tabindex=-1>out</span>
<div focusgroup>
  <span id=item1 tabindex=0>item1</span>
  <span id=item2 tabindex=-1>item2</span>
</div>

<script>

  promise_test(async t => {
    var out = document.getElementById("out");

    await focusAndKeyPress(out, kArrowDown);
    assert_equals(document.activeElement, out);

    await focusAndKeyPress(out, kArrowRight);
    assert_equals(document.activeElement, out);
  }, "When the focus is set on an element outside of the focusgroup, an arrow keypress shouldn't move the focus at all.");

</script>
